Added instructions for customizing .ariax
Unfortunately the default .ariax presets (preset groups for ARIA) use absolute paths, and the ariax format does not allow for relative paths except for "official libraries". As a result, each user will need to change the path for every 'slot' in the ariax files they plan on using until a workaround can be found.